@@ -15,51 +20,50 @@
   },
   "scripts": {
     "pretest": "npm run lint",
-    "test": "npm run vscode:test",
-    "vscode:test": "node --require=@adonisjs/require-ts/build/register bin/test.ts",
-    "build": "npm run compile",
+    "test": "c8 npm run quick:test",
     "clean": "del-cli build",
-    "compile": "npm run lint && npm run clean && tsc",
-    "release": "np",
-    "version": "npm run build",
+    "typecheck": "tsc --noEmit",
+    "precompile": "npm run lint && npm run clean",
+    "compile": "tsup-node && tsc --emitDeclarationOnly --declaration",
+    "build": "npm run compile",
     "prepublishOnly": "npm run build",
     "lint": "eslint . --ext=.ts",
+    "format": "prettier --write .",
+    "release": "np",
+    "version": "npm run build",
     "sync-labels": "github-label-sync --labels .github/labels.json edge-js/lexer",
-    "format": "prettier --write ."
+    "quick:test": "node --enable-source-maps --loader=ts-node/esm bin/test.ts"
   },
-  "keywords": [
-    "edge",
-    "template",
-    "template-engine"
-  ],
-  "author": "virk",
-  "license": "MIT",
   "devDependencies": {
-    "@adonisjs/require-ts": "^2.0.13",
-    "@commitlint/cli": "^17.4.4",
-    "@commitlint/config-conventional": "^17.4.4",
-    "@japa/assert": "^1.4.1",
-    "@japa/run-failed-tests": "^1.1.1",
-    "@japa/runner": "^2.5.1",
-    "@japa/spec-reporter": "^1.3.3",
-    "@types/dedent": "^0.7.0",
-    "@types/node": "^18.15.3",
+    "@adonisjs/eslint-config": "^1.1.8",
+    "@adonisjs/prettier-config": "^1.1.8",
+    "@adonisjs/tsconfig": "^1.1.8",
+    "@commitlint/cli": "^18.2.0",
+    "@commitlint/config-conventional": "^18.1.0",
+    "@japa/assert": "^2.0.1",
+    "@japa/runner": "^3.0.5",
+    "@swc/core": "^1.3.96",
+    "@types/dedent": "^0.7.1",
+    "@types/node": "^20.8.10",
     "benchmark": "^2.1.4",
-    "dedent": "^0.7.0",
-    "del-cli": "^5.0.0",
-    "eslint": "^8.36.0",
-    "eslint-config-prettier": "^8.7.0",
-    "eslint-plugin-adonis": "^2.1.1",
-    "eslint-plugin-prettier": "^4.2.1",
+    "c8": "^8.0.0",
+    "dedent": "^1.5.1",
+    "del-cli": "^5.1.0",
+    "eslint": "^8.53.0",
     "github-label-sync": "^2.3.1",
     "husky": "^8.0.3",
-    "np": "^7.6.3",
-    "prettier": "^2.8.5",
-    "typescript": "^5.0.2"
+    "np": "^8.0.4",
+    "prettier": "^3.0.3",
+    "ts-node": "^10.9.1",
+    "tsup": "^7.1.0",
+    "typescript": "^5.2.2"
   },
   "dependencies": {
-    "edge-error": "^3.0.0"
+    "edge-error": "^4.0.0"
   },
+  "author": "virk",
+  "license": "MIT",
+  "homepage": "https://github.com/edge-js/lexer#readme",
   "repository": {
     "type": "git",
     "url": "git+https://github.com/edge-js/lexer.git"
