@@ -7,27 +7,28 @@
  * file that was distributed with this source code.
  */
 
-import { test } from '@japa/runner'
 import dedent from 'dedent'
-import { Tokenizer } from '../src/tokenizer'
-import { TagTypes, MustacheTypes } from '../src/types'
+import { test } from '@japa/runner'
+
+import { Tokenizer } from '../src/tokenizer.js'
+import { TagTypes, MustacheTypes } from '../src/enums.js'
 
 const tagsDef = {
   if: class If {
-    public static block = true
-    public static seekable = true
+    static block = true
+    static seekable = true
   },
   else: class Else {
-    public static block = false
-    public static seekable = false
+    static block = false
+    static seekable = false
   },
   include: class Include {
-    public static block = false
-    public static seekable = true
+    static block = false
+    static seekable = true
   },
   each: class Each {
-    public static block = true
-    public static seekable = true
+    static block = true
+    static seekable = true
   },
 }
 
@@ -1062,9 +1063,9 @@ test.group('Tokenizer | Tags', () => {
 
     const tags = {
       if: class If {
-        public static block = true
-        public static seekable = true
-        public static noNewLine = true
+        static block = true
+        static seekable = true
+        static noNewLine = true
       },
     }
 
@@ -1135,9 +1136,9 @@ test.group('Tokenizer | Tags', () => {
 
     const tags = {
       if: class If {
-        public static block = true
-        public static seekable = true
-        public static noNewLine = true
+        static block = true
+        static seekable = true
+        static noNewLine = true
       },
     }
 
@@ -1259,6 +1260,61 @@ test.group('Tokenizer | Tags', () => {
         filename: 'eval.edge',
         value: 'world',
         line: 4,
+      },
+    ])
+  })
+
+  test('remove newline with non-seekable tag', ({ assert }) => {
+    const template = dedent`
+    Hello
+    @include~
+    Hi
+    `
+
+    const tags = {
+      include: class Include {
+        static block = false
+        static seekable = false
+        static noNewLine = false
+      },
+    }
+
+    const tokenizer = new Tokenizer(template, tags, { filename: 'eval.edge' })
+    tokenizer.parse()
+
+    assert.isNull(tokenizer.tagStatement)
+    assert.deepEqual(tokenizer.tokens, [
+      {
+        type: 'raw',
+        filename: 'eval.edge',
+        value: 'Hello',
+        line: 1,
+      },
+      {
+        type: TagTypes.TAG,
+        filename: 'eval.edge',
+        children: [],
+        loc: {
+          start: {
+            line: 2,
+            col: 8,
+          },
+          end: {
+            line: 2,
+            col: 8,
+          },
+        },
+        properties: {
+          name: 'include',
+          jsArg: '',
+          selfclosed: false,
+        },
+      },
+      {
+        type: 'raw',
+        filename: 'eval.edge',
+        value: 'Hi',
+        line: 3,
       },
     ])
   })
